Rename package for FIDO2 SDK
