+++ sdk/fido2/src/main/java/com/ibm/security/verifysdk/fido2/BiometricAuthenticationException.kt
@@ -2,7 +2,7 @@
  *  Copyright contributors to the IBM Security Verify FIDO2 SDK for Android project
  */
 
-package com.ibm.security.verify.fido2
+package com.ibm.security.verifysdk.fido2
 
 /**
  * Represents an exception specific to biometric authentication.
