+++ sdk/fido2/src/main/java/com/ibm/security/verifysdk/fido2/model/AssertionOptions.kt
@@ -1,7 +1,7 @@
 /*
  *  Copyright contributors to the IBM Security Verify FIDO2 SDK for Android project
  */
-package com.ibm.security.verify.fido2.model
+package com.ibm.security.verifysdk.fido2.model
 
 import kotlinx.serialization.Serializable
