+++ sdk/fido2/src/main/java/com/ibm/security/verifysdk/fido2/model/ClientDataJsonAttestation.kt
@@ -1,20 +1,22 @@
 /*
  *  Copyright contributors to the IBM Security Verify FIDO2 SDK for Android project
  */
-package com.ibm.security.verify.fido2.model
+package com.ibm.security.verifysdk.fido2.model
 
 import kotlinx.serialization.Serializable
 
 /**
- * Represents the client data for assertion during a WebAuthn operation.
+ * Represents the client data for attestation during a WebAuthn operation.
  *
  * @property type The type of the client data.
  * @property challenge The challenge associated with the client data.
  * @property origin The origin associated with the client data.
+ * @property crossOrigin Indicates whether the client data is from a cross-origin operation.
  */
 @Serializable
-data class ClientDataJsonAssertion(
+data class ClientDataJsonAttestation(
     val type: String,
     val challenge: String,
     val origin: String,
+    val crossOrigin: Boolean
 )
