+++ sdk/fido2/src/test/java/com/ibm/security/verifysdk/fido2/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.ibm.security.verify.fido2
+package com.ibm.security.verifysdk.fido2
 
 import org.junit.Test
